ContextSupport improved using ContextSupportPreferences
